update confirmation box

+++ FrontEnd/src/app/service/api.service.ts
@@ -94,7 +94,17 @@ export class ApiService {
   }
 
   remove_userRequest(orderid:any){
-    return this.http.delete<any>(`${environment.SERVER_URI}/api/user-request/delete/${orderid}`)
-    .subscribe()
+    var result: any;
+    if (window.confirm(`Deleting your requested Order# ${orderid}`)){
+      try{
+        result = this.http.delete<any>(`${environment.SERVER_URI}/api/user-request/delete/${orderid}`)
+        .subscribe()
+      }
+      catch (err) {
+        alert(err)
+      }
+    }
+    return result
   }
+
 }
